pyproject.toml
[tool.poetry]
name = "test02"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
graia-ariadne = "^0.6.10"
graia-saya = "^0.0.15"
graia-scheduler = "^0.0.7"
arclet-alconna = "^0.7.5"
uvicorn = "^0.17.6"
fastapi = "^0.75.0"
graia-broadcast = "^0.16.1"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
# 这里以清华源举例，你也可以使用其他源
name = "tuna-tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = false